pyproject.toml
[project]
name = "xilogism"
version = "0.8.7"
description = "A companion application for Computer Engineering Students taking Logic Circuits and Design"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "treppenwitz03", email = "owendavidmalicsi@gmail.com" }
]
dependencies = [
  "flet==0.28.2",
  "xilowidgets @ git+https://github.com/logihelpers/xilowidgets.git",
  "flet-layoutbuilder @ git+https://github.com/Wanna-Pizza/flet_layoutbuilder",
  "python-docx",
  "docx2pdf",
  "pypdfium2",
  "pillow",
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "io.logihelpers"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "xilogism"

# company name to display in about app dialogs
company = "Logihelpers"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by Logihelpers"

[tool.flet.app]
path = "src"

[tool.uv]
dev-dependencies = [
    "flet[all]==0.28.2",
    "docx2pdf==0.1.8",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.2"}